Move rpc to core, rpc.cluster to cluster
