+++ grakn/core/stream.py
