+++ grakn/core/utils.py
