+++ tests/behaviour/connection/session/session_steps.py
@@ -24,7 +24,7 @@ from typing import List
 from behave import *
 from hamcrest import *
 
-from grakn.rpc.session import SessionType
+from grakn.core.session import SessionType
 from tests.behaviour.config.parameters import parse_bool, parse_list
 from tests.behaviour.context import Context
